move method to RDS instance struct
